fix some deps
