+++ app/client/src/PluginActionEditor/components/PluginActionResponse/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as hasFailed } from "./actionHasFailed";
+export { default as isHtml } from "./isHtml";
