+++ app/client/src/PluginActionEditor/components/PluginActionResponse/utils/isHtml.ts
@@ -1,0 +1,9 @@
+const isHtml = (str: string) => {
+  const doc = new DOMParser().parseFromString(str, "text/html");
+
+  return Array.from(doc.body.childNodes).some(
+    (node: ChildNode) => node.nodeType === 1,
+  );
+};
+
+export default isHtml;
