+++ app/client/src/components/editorComponents/utils.ts
@@ -1,15 +1,13 @@
 import type { JSAction } from "entities/JSCollection";
-import { JSResponseState } from "./JSResponseView";
 
-export const isHtml = (str: string) => {
-  const doc = new DOMParser().parseFromString(str, "text/html");
-
-  return Array.from(doc.body.childNodes).some(
-    // TODO: Fix this the next time the file is edited
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (node: any) => node.nodeType === 1,
-  );
-};
+export enum JSResponseState {
+  IsExecuting = "IsExecuting",
+  IsDirty = "IsDirty",
+  IsUpdating = "IsUpdating",
+  NoResponse = "NoResponse",
+  ShowResponse = "ShowResponse",
+  NoReturnValue = "NoReturnValue",
+}
 
 /**
  * Returns state of the JSResponseview editor component
